+++ wemake_python_styleguide/violations/base.py
@@ -76,7 +76,8 @@ class BaseViolation(object):
         self._node = node
 
         if text is None:
-            self._text = node.__class__.__name__.lower()
+            # TODO: why do we need this line?
+            self._text = node.__class__.__qualname__.lower()
         else:
             self._text = text