+++ CHANGELOG.md
@@ -18,6 +18,7 @@ We used to have incremental versioning before `0.1.0`.
 - Forbids to subclass `BaseException`
 - Forbids to use simplifiable `if` expressions
 - Forbids to use simplifiable `if` nodes
+- Forbids to have incorrect nodes in `class` body
 
 ### Bugfixes